+++ src/main/java/kitchenpos/products/infra/DefaultPurgomalumClient.java
@@ -1,5 +1,6 @@
 package kitchenpos.products.infra;
 
+import kitchenpos.products.domain.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
